upgrade eth-beacon to v3
