--- .changeset/eighty-lemons-unite.md
@@ -1,5 +1,0 @@
----
-'@chainlink/eth-beacon-test-adapter': major
----
-
-Initial version of EA
