+++ .changeset/three-spiders-flash.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/eth-beacon-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
