--- packages/sources/eth-beacon-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as balance } from './balance'
