--- packages/sources/eth-beacon-test/test-payload.json
@@ -1,10 +1,0 @@
-{
- "requests": [{
-  "result": [
-   {
-    "address":
-    "0x8bdb63ea991f42129d6defa8d3cc5926108232c89824ad50d57f49a0310de73e81e491eae6587bd1465fa5fd8e4dee21",
-   }
-  ]
- }]
-}
