+++ packages/sources/eth-beacon/package.json
@@ -27,20 +27,16 @@
     "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
     "start": "yarn server:dist"
   },
-  "dependencies": {
-    "@chainlink/ea-bootstrap": "workspace:*",
-    "@chainlink/ea-test-helpers": "workspace:*",
-    "axios": "1.3.6",
-    "bignumber.js": "9.1.1",
-    "ethers": "5.7.2",
-    "tslib": "^2.3.1"
-  },
   "devDependencies": {
     "@types/jest": "27.5.2",
-    "@types/node": "16.11.47",
-    "@types/supertest": "2.0.12",
+    "@types/node": "16.11.51",
     "nock": "13.2.9",
-    "supertest": "6.2.4",
     "typescript": "5.0.4"
+  },
+  "dependencies": {
+    "@chainlink/external-adapter-framework": "0.31.0",
+    "bignumber.js": "9.1.1",
+    "ethers": "5.7.2",
+    "tslib": "2.4.1"
   }
 }
