--- packages/sources/eth-beacon/schemas/env.json
@@ -1,40 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/example-source-adapter.json",
-  "title": "Chainlink External Adapter for Ethereum Beacon API",
-  "description": "External adapter for reading from the Ethereum PoS Beacon chain's API and optionally searching deposit events for limbo validator balances.",
-  "required": ["ETH_CONSENSUS_RPC_URL"],
-  "type": "object",
-  "properties": {
-    "ETH_CONSENSUS_RPC_URL": {
-      "type": "string",
-      "description": "RPC URL of an Ethereum consensus client (beacon node)",
-      "required": true
-    },
-    "ETH_EXECUTION_RPC_URL": {
-      "type": "string",
-      "description": "RPC URL of an Ethereum execution client (archive node). Required for requests that need a limbo validator search",
-      "required": false
-    },
-    "BATCH_SIZE": {
-      "type": "number",
-      "description": "Number of validators to send in each request to the consensus client. Set to 0 if consensus client allows unlimited validators in query. Setting this lower than the default and greater than 0 may result in lower performance from the adapter.",
-      "default": "15"
-    },
-    "GROUP_SIZE": {
-      "type": "number",
-      "description": "Number of requests to execute asynchronously before the adapter waits to execute the next group of requests. Setting this lower than the default may result in lower performance from the adapter. Unused if BATCH_SIZE is set to 0.",
-      "default": "15"
-    },
-    "CHAIN_ID": {
-      "required": false,
-      "description": "The chain id to connect to",
-      "type": "string",
-      "default": "1"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
