--- packages/sources/eth-beacon/src/abi/DepositAbi.ts
@@ -1,34 +1,0 @@
-export const DepositEvent_ABI = [
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: 'bytes',
-        name: 'pubkey',
-        type: 'bytes',
-      },
-      {
-        indexed: false,
-        internalType: 'bytes',
-        name: 'withdrawal_credentials',
-        type: 'bytes',
-      },
-      {
-        indexed: false,
-        internalType: 'bytes',
-        name: 'amount',
-        type: 'bytes',
-      },
-      {
-        indexed: false,
-        internalType: 'bytes',
-        name: 'signature',
-        type: 'bytes',
-      },
-      { indexed: false, internalType: 'bytes', name: 'index', type: 'bytes' },
-    ],
-    name: 'DepositEvent',
-    type: 'event',
-  },
-]
