+++ packages/sources/eth-beacon/src/config/DepositAbi.ts
