--- packages/sources/eth-beacon/src/config/envDefaultOverrides.ts
@@ -1,5 +1,0 @@
-import type { EnvDefaultOverrides } from '@chainlink/ea-bootstrap'
-
-export const envDefaultOverrides: EnvDefaultOverrides = {
-  API_TIMEOUT: '60000',
-}
