+++ packages/sources/eth-beacon/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as BalanceInputParameters } from './balance'
-
-export type TInputParameters = BalanceInputParameters
-
-export * as balance from './balance'
+export { endpoint as balance } from './balance'
