+++ packages/sources/eth-beacon/src/transport/utils.ts
