+++ packages/sources/eth-beacon/test/integration/fixtures.ts
@@ -2,6 +2,7 @@ import nock from 'nock'
 
 export const mockBalanceSuccess = (): nock.Scope =>
   nock('http://localhost:3500', { encodedQueryParams: true })
+    .persist()
     .get(
       '/eth/v1/beacon/states/finalized/validators?id=0x8bdb63ea991f42129d6defa8d3cc5926108232c89824ad50d57f49a0310de73e81e491eae6587bd1465fa5fd8e4dee21,0xb672b5976879c6423ad484ba4fa0e76069684eed8e2a8081f6730907f3618d43828d1b399d2fd22d7961824594f73462',
     )
@@ -57,6 +58,7 @@ export const mockBalanceSuccess = (): nock.Scope =>
         'Wed, 21 Sep 2022 10:58:34 GMT',
       ],
     )
+    .persist()
 
 export const mockBalanceWithStatusSuccess = (): void => {
   nock('http://localhost:3500', { encodedQueryParams: true })
@@ -186,4 +188,5 @@ export const mockBalanceLimboValidator = (): void => {
         'Wed, 21 Sep 2022 10:58:34 GMT',
       ],
     )
+    .persist()
 }
