--- packages/sources/eth-beacon/test/unit/README.md
@@ -1,1 +1,0 @@
-Tests that do not run the full adapter, instead testing smaller pieces such as endpoints or helper functions.
