+++ packages/tsconfig.test.json
@@ -303,9 +303,6 @@
       "path": "./sources/eth-beacon/tsconfig.test.json"
     },
     {
-      "path": "./sources/eth-beacon-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/etherchain/tsconfig.test.json"
     },
     {
